This Is CodeAlpha 3rd project.
📚 Personal Book Library Web App
Welcome to the Personal Book Library Web App! This project was developed during my 3-month remote internship at CodeAlpha. It is a simple yet powerful tool to manage your personal book collection.

🌟 Features
🔍 Book Search: Easily search for books in your collection by title, author, or category.
📚 Categorization: Organize your books into different categories for better management.
💾 Local Storage: Your book data is securely stored in the browser's local storage, ensuring your collection is saved even after you close the browser.

🛠️ Technologies Used
HTML5 for the structure of the web app.
CSS3 for styling the web app.
JavaScript for adding functionality to the web app.
Local Storage for saving book data in the browser.